Darker navbar & footer, button icons, episode listing, about page
The footer and navbar received a darker background color to make the text inside them more readable.
The Typicons icon font was added to the project. Some buttons already received their icons.
The /episodes page displays all episodes in the database, but the pages where it links to do not exist yet.
The /about page's last section was replaced with something more informative, and the page also got a logo and title on the top.